cast

+++ cleverhans/attacks_tf.py
@@ -1742,8 +1742,10 @@ def _project_perturbation(perturbation, epsilon, input_image, clip_min=None,
 
   # Ensure inputs are in the correct range
   with tf.control_dependencies([
-      utils_tf.assert_less_equal(input_image, clip_max),
-      utils_tf.assert_greater_equal(input_image, clip_min)
+      utils_tf.assert_less_equal(input_image,
+                                 tf.cast(clip_max, input_image.dtype)),
+      utils_tf.assert_greater_equal(input_image,
+                                    tf.cast(clip_min, input_image.dtype))
   ]):
     clipped_perturbation = utils_tf.clip_by_value(
         perturbation, -epsilon, epsilon)
